llm-review-service/src/axon-sidecar/requirements.txt
axoniq>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
pydantic>=2.10.0
gitpython>=3.1.44

# dev/test
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-httpx>=0.35.0
